requirements.txt
aniso8601==9.0.1
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==1.4.6
passlib==1.7.4
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
